Añado condicion if

+++ Jenkinsfile
@@ -13,7 +13,9 @@ pipeline {
             steps {
                 catchError(buildResult: 'UNSTABLE', stageResult: 'FAILURE'){
                     sh '''
-                        rm flake8.out
+                        if [ -f flake8.out ]; then
+                            rm flake8.out
+                        fi
                         sudo python3 -m flake8 --exit-zero --format=pylint app >flake8.out
                     '''
                     recordIssues tools: [flake8(name: 'Flake8', pattern: 'flake8.out')]
@@ -24,7 +26,9 @@ pipeline {
             steps {
                 catchError(buildResult: 'UNSTABLE', stageResult: 'FAILURE'){
                      sh '''
-                        rm bandit.out
+                        if [ -f bandit.out ]; then
+                            rm bandit.out
+                        fi
                         sudo python3 -m bandit -r . -f custom -o bandit.out -ll -ii --msg-template "{abspath}:{line}: [{test_id}] {msg}"
                     '''
                     recordIssues tools: [pyLint(name: 'Bandit', pattern: 'bandit.out')]
